% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCustomOutput.R
\name{createCustomOutput}
\alias{createCustomOutput}
\title{Create custom output for this study}
\usage{
createCustomOutput(outputFolder)
}
\arguments{
\item{outputFolder}{The path to the output folder containing the results.}
}
\description{
Create custom output for this study
}
\details{
Performs empricical calibration using the negative control outcomes, and computes the calibrated
p-values
}

