% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the novel oral anticoagulant study}
\usage{
execute(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema, studyCohortTable = "cohort_noac",
  studyCohortDefinitionTable = "cohort_definition_noac",
  oracleTempSchema = NULL, cdmVersion = 5, outputFolder,
  drugTarget = c("Rivaroxaban", "Dabigatran"), analysisDesign = "main",
  createCohorts = TRUE, runAnalyses = TRUE, packageResults = TRUE,
  maxCores = 4)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using
the \code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format
resides. Note that for SQL Server, this should include both
the database and schema name, for example 'cdm_data.dbo'.}

\item{workDatabaseSchema}{Schema name where intermediate data can be stored. You will
need to have write priviliges in this schema. Note that for
SQL Server, this should include both the database and schema
name, for example 'cdm_data.dbo'.}

\item{studyCohortTable}{The name of the table that will be created in the work
database schema. This table will hold the exposure and outcome
cohorts used in this study.}

\item{studyCohortDefinitionTable}{The table in the work database schema containing the
definitions of the cohorts.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user
has write priviliges for storing temporary tables.}

\item{cdmVersion}{Version of the CDM. Can be "4" or "5"}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/)}

\item{analysisDesign}{Select which analysis designs to run: `main` or `all`}

\item{createCohorts}{Create cohort tables from database (TRUE) or use previously
saved cohort information (FALSE)}

\item{runAnalyses}{Run propensity score and outcome model fitting (TRUE) or use
previously saved models (FALSE)}

\item{packageResults}{Package the results for sharing?}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{targetDrug}{Select which target drugs to run: `Rivaroxaban`, `Dabigatran`
or both}
}
\value{
Nothing.  All intermediate files, models and reports are in `outputFolder`.
}
\description{
This function executes the novel oral anticoagulant study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "mdcr_v5",
        workDatabaseSchema = "ohdsi",
        outputFolder = "~/study_results",
       maxCores = 4)

}

}

